strlcat corr

+++ lib/ft_strlcat.c
@@ -6,16 +6,18 @@
 /*   By: sscarecr <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/05 18:22:50 by sscarecr          #+#    #+#             */
-/*   Updated: 2019/09/05 18:24:04 by sscarecr         ###   ########.fr       */
+/*   Updated: 2019/09/05 18:54:47 by sscarecr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
+#include <stdlib.h>
+
+size_t ft_strlcat(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	j;
-	unsigned int	slen;
-	unsigned int	dlen;
+	size_t	i;
+	size_t	j;
+	size_t	slen;
+	size_t	dlen;
 
 	dlen = 0;
 	slen = 0;
